Fix for PipeReader contents feching.
Ignore-this: d3362cb2b00c407b9ccc6adc97fdbf0d

Apparently, a pipe has to be opened in r/w mode for xmobar to update the
display of its contents when the latter changes in time. I don't see the
reason for this, but the fact remains that opening the pipe in ReadMode does
not work.

darcs-hash:20090926215132-a9375-0b3384071129d6352493c959fbf078e3a9f58709.gz