Cope with new file system layout of CPU temperature sensor readings
Either due to a change in lm_sensors or in the kernel itself, the core
temperature readings are no longer found in one directory per core but in
multiple files in the same directory.  What's worse is that the file names have
little to do with the actual core number.  This patch addresses this and is
flexible enough to handle the old layout and the new one.