Fix the detection of the end of the command array
It was only looked for a ']' followed by space-characters and either a
'}' or ','.

This looks good in generally but considering that a command can have as
its last input parameter an array e.g.

...
  , Run Something ["Do"]
  , Run XMonadLog
  ]
, sepChar = "%"
}
...

This will not work unless 'Something' is the last command in the array.

This patch fixes this problem, as a ',' is only accepted if it is not
followed by a "Run".