Working version
The last commit removed the exposure event which turned out to be a big
problem.

Although the bug still exists that not all xrandr events are received
when normal events are enabled.

To work around this problem a second display is created on which only
the xrandr events are enabled.
On an exposure event the eventqueue for this display is processed.

The results are very good, in the worst case an exposure event must be
triggered by the user on xmobar to update its position.