removed threading problem with Xlib
The output just stopped at some point until a new XEvent was received

As XLockDisplay is in theory a good idea, with XNextEvent blocking its
not usable.

As it turned out, a window can be shared between display connections.

Now the eventloop has its own display connection (which also removes the
need for the lock introduced before).

Additionally the screeninfo doesn't need to be fetched into a TVar in
the eventerloop anymore.

Also this was needed for the signalHandlers to work correctly again.