DateZone, Date*L Plugin merges; DateZone API-Change!
To support multiple locales, an lock was introduced.
Although through this, supporting DateL and DateZoneL would be
cumbersome.

To simplify the usage, DateZone was replaced with DateZoneL.
Additionally the position of the Alias parameter was changed.

Using "" as Zone parameter for DateZoneL simulates DateL.

Providing also "" for locale simulates Date.