Add functions for {reveal,hid,toggl}ing the window
Toggling is based is based on the current window status. If unmapped
then reveal else hide.

Sync is necessary or delays might occur.

The functions are called from the event loop when the according signal
is received

When mapping (revealing) the window again we need to set the struts
property again. The easiest way to do this is to call repositionWin.
However, repositionWin needs access to the Config structure which is
available in eventLoop.

Because decomposition wouldn't be easy and I don't want to pass Config
down to showWindow (which would need to return the new Rectangle then)
this is done here.