Allow user to define punctuation.
I'd like to be able to choose which punctuation to use. All my other xmobar plugins use a colon and a space (e.g. `B: 100%` for battery), but this one leaves the space out (e.g. `E:1`) which is mildly annoying. It would be nice to let the user choose if he/she would like to use something other than a colon or to add extra whitespace.