Guard monitor 'Weather' behind flag 'with_weather'
Hide the monitor 'Weather' behind a flag named 'with_weather'.
This implies that the dependency on the module 'HTTP' is now optional.
The default for 'with_weather' is True since we do not want to break
backwards compatibility/user experience.

Further moved the effect of the flag with_conduit within the branch of
'with_weather', since that flag only has some effect if 'with_weather'
is true.