Fix UVMeter due to changed URL and BOM
The URL which exports the real time UV data has changed to an HTTPS
address. Since the HTTP package does not support HTTPS URLs, use
http-conduit to retrieve the XML document.

Unfortunately, the XML documents XML declaration precedes a byte order
mark which the previous XML parser was unable to handle. We're simply
ignoring the BOM in order to get to the UV values.