Implement timer coalescing (noticeably less CPU/power usage)
xmobar currently runs every monitor in its own thread. Monitors that do
periodic updates simply sleep and loop. This unfortunately leads to
these threads coming out of sync, and xmobar ends up waking up and
redrawing for every periodic monitor. In my case, that is 7 times per
second, which is enough for xmobar to be at the top of "top" with more
than 1% CPU usage, and to have a noticeable impact on battery life.

This commit adds a central timer coordination thread which makes sure
that periodic updates happen together and that we only redraw once
they're all done.

Together with PR #409, I managed to lower the idle power draw of my
laptop from 4W to 3W.