Timer coalescing: gracefully uncoalesce slow timers
The first implementation assumed all timers (monitors) are fast and
frequent (which happens to be the case in my configuration). This meant
that a single on-line weather monitor could block the entire xmobar
instance for a long time due to the refresh pausing (meant to reduce
power consumption).

This commit attempts to fix that by limiting the refresh pause time and
using the old periodic sleep method for these slow timers (monitors).