Add k10temp plugin
The existing support for the coretemp kernel driver only works with
Intel CPUs.

This commit extends support for temperature monitoring to AMD CPUs.

k10temp is a kernel driver for monitoring the temperature of AMD
processors. It supports everything from AMD's 10h (Opteron/Athlon)
family of processors to the latest 19h (Zen 3) family.

Reference: https://www.kernel.org/doc/html/latest/hwmon/k10temp.html

The meaning of the various temperatures made available has changed over
the years and only `Tctl` is available on processors prior to the 17h
family.

Labels for these temperatures are present but as Tctl and Tdie do not
contain a number I could not find a way to use these as
`checkedDataRetrieval` expects an integer label.

It is a PCI device and so an address needs to be supplied as part of the
configuration.

Example configuration:
`Run K10Temp "0000:00:18.3" ["--template", "T: <Tdie>C | <Tccd1>C"] 60`